chat_bot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "groq>=0.9.0",
    "mem0ai>=0.1.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.2.3",
]
